[BUILD] Fix GCC warnings
Problem: Signedness, macro and unused-code warnings.

Why: Implicit conversions, ungrouped operands and dead code.

How: Make types and grouping explicit; remove unused code.

+++ dll/win32/devmgr/devmgmt/DeviceView.cpp
@@ -711,7 +711,7 @@ CDeviceView::AddResourceNodes(
         }
         else if (Descriptor->Type == CmResourceTypeMemory)
         {
-            hParent = (Descriptor->u.Memory.Start.QuadPart >= 0x100000000ULL) ? Items->hLargeMemory : Items->hMemory;
+            hParent = ((ULONGLONG)Descriptor->u.Memory.Start.QuadPart >= 0x100000000ULL) ? Items->hLargeMemory : Items->hMemory;
         }
         else if (Descriptor->Type == CmResourceTypeDma)
         {
